start on refactor
